[#12] Floor → Room → Start 예외 처리
## 수정 사항
- Floor 선택시 Room을 ''로 초기화 해서 Floor → Start flow를 막음
