+++ src/assets/F1Rooms.ts
@@ -1,0 +1,18 @@
+const F1Rooms: string[] = [
+  '101',
+  '102',
+  '103',
+  '104',
+  '105',
+  '106',
+  '107',
+  '108',
+  '109',
+  '110',
+  '111',
+  '112',
+  '113',
+  '114',
+];
+
+export default F1Rooms;
