+++ src/assets/F2Rooms.ts
@@ -1,0 +1,20 @@
+const F2Rooms: string[] = [
+  '201',
+  '202',
+  '203',
+  '204',
+  '205',
+  '206',
+  '207',
+  '208',
+  '209',
+  '210',
+  '211',
+  '212',
+  '213',
+  '214',
+  '215',
+  '215-1',
+];
+
+export default F2Rooms;
