+++ src/assets/F3Rooms.ts
@@ -1,0 +1,24 @@
+const F3Rooms: string[] = [
+  '301',
+  '302',
+  '303',
+  '304',
+  '305',
+  '306',
+  '307',
+  '308',
+  '309',
+  '310',
+  '311',
+  '312',
+  '313',
+  '314',
+  '315',
+  '316',
+  '317',
+  '318',
+  '319',
+  '320',
+];
+
+export default F3Rooms;
