+++ src/assets/F4Rooms.ts
@@ -1,0 +1,20 @@
+const F4Rooms: string[] = [
+  '401',
+  '402',
+  '403',
+  '404',
+  '405',
+  '406',
+  '407',
+  '408',
+  '409',
+  '410',
+  '411',
+  '412',
+  '413',
+  '414',
+  '415',
+  '416',
+];
+
+export default F4Rooms;
