+++ src/assets/F5Rooms.ts
@@ -1,0 +1,37 @@
+const F5Rooms: string[] = [
+  '501',
+  '502',
+  '503',
+  '504',
+  '505',
+  '506',
+  '507',
+  '508',
+  '509',
+  '510',
+  '511',
+  '512',
+  '513',
+  '514',
+  '515',
+  '516',
+  '517',
+  '518',
+  '519',
+  '520',
+  '521',
+  '522',
+  '523',
+  '524',
+  '525',
+  '526',
+  '527',
+  '528',
+  '529',
+  '530',
+  '531',
+  '532',
+  '533',
+];
+
+export default F5Rooms;
