+++ src/assets/F6Rooms.ts
@@ -1,0 +1,37 @@
+const F6Rooms: string[] = [
+  '601',
+  '602',
+  '603',
+  '604',
+  '605',
+  '606',
+  '607',
+  '608',
+  '609',
+  '610',
+  '611',
+  '612',
+  '613',
+  '614',
+  '615',
+  '616',
+  '617',
+  '618',
+  '619',
+  '620',
+  '621',
+  '622',
+  '623',
+  '624',
+  '625',
+  '626',
+  '627',
+  '628',
+  '629',
+  '630',
+  '631',
+  '632',
+  '633',
+];
+
+export default F6Rooms;
