+++ src/components/SpeedTestPage/components/LocateFloorRoom.tsx
@@ -1,5 +1,5 @@
 /* RecoilHook */
-import { useRecoilState } from 'recoil';
+import { useRecoilState, useSetRecoilState } from 'recoil';
 
 /* MUI */
 import {
@@ -14,6 +14,13 @@ import Select, { SelectChangeEvent } from '@mui/material/Select';
 
 /* Atom */
 import { floorState, roomState } from '../../../module/Atom';
+
+import F1Rooms from '../../../assets/F1Rooms';
+import F2Rooms from '../../../assets/F2Rooms';
+import F3Rooms from '../../../assets/F3Rooms';
+import F4Rooms from '../../../assets/F4Rooms';
+import F5Rooms from '../../../assets/F5Rooms';
+import F6Rooms from '../../../assets/F6Rooms';
 
 const SelectionTheme = createTheme({
   typography: {
@@ -60,156 +67,6 @@ const RoomsValid = ({
 };
 
 const RoomSplit = ({ floor }: { floor: string }) => {
-  const F1Rooms: string[] = [
-    '101',
-    '102',
-    '103',
-    '104',
-    '105',
-    '106',
-    '107',
-    '108',
-    '109',
-    '110',
-    '111',
-    '112',
-    '113',
-    '114',
-  ];
-
-  const F2Rooms: string[] = [
-    '201',
-    '202',
-    '203',
-    '204',
-    '205',
-    '206',
-    '207',
-    '208',
-    '209',
-    '210',
-    '211',
-    '212',
-    '213',
-    '214',
-    '215',
-    '215-1',
-  ];
-
-  const F3Rooms: string[] = [
-    '301',
-    '302',
-    '303',
-    '304',
-    '305',
-    '306',
-    '307',
-    '308',
-    '309',
-    '310',
-    '311',
-    '312',
-    '313',
-    '314',
-    '315',
-    '316',
-    '317',
-    '318',
-    '319',
-    '320',
-  ];
-
-  const F4Rooms: string[] = [
-    '401',
-    '402',
-    '403',
-    '404',
-    '405',
-    '406',
-    '407',
-    '408',
-    '409',
-    '410',
-    '411',
-    '412',
-    '413',
-    '414',
-    '415',
-    '416',
-  ];
-
-  const F5Rooms: string[] = [
-    '501',
-    '502',
-    '503',
-    '504',
-    '505',
-    '506',
-    '507',
-    '508',
-    '509',
-    '510',
-    '511',
-    '512',
-    '513',
-    '514',
-    '515',
-    '516',
-    '517',
-    '518',
-    '519',
-    '520',
-    '521',
-    '522',
-    '523',
-    '524',
-    '525',
-    '526',
-    '527',
-    '528',
-    '529',
-    '530',
-    '531',
-    '532',
-    '533',
-  ];
-
-  const F6Rooms: string[] = [
-    '601',
-    '602',
-    '603',
-    '604',
-    '605',
-    '606',
-    '607',
-    '608',
-    '609',
-    '610',
-    '611',
-    '612',
-    '613',
-    '614',
-    '615',
-    '616',
-    '617',
-    '618',
-    '619',
-    '620',
-    '621',
-    '622',
-    '623',
-    '624',
-    '625',
-    '626',
-    '627',
-    '628',
-    '629',
-    '630',
-    '631',
-    '632',
-    '633',
-  ];
-
   switch (floor) {
     case '1':
       return (
@@ -258,8 +115,10 @@ const RoomSplit = ({ floor }: { floor: string }) => {
 
 const LocateFloorRoom = () => {
   const [floor, setFloor] = useRecoilState(floorState);
+  const setRoom = useSetRecoilState(roomState);
   const floorChange = (event: SelectChangeEvent) => {
     setFloor(event.target.value as string);
+    setRoom('');
   };
 
   const floorArray: string[] = ['1', '2', '3', '4', '5', '6'];
